naming update and feature added
